requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-extensions==2.2.9
factory-boy==2.12.0
psycopg2==2.8.5
